added confirm before deleting estate functionality

+++ client/src/components/Estate/Estate.jsx
@@ -1,14 +1,33 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import './Estate.scss';
 import {connect} from 'react-redux';
 import {Link, withRouter} from 'react-router-dom';
 import {deleteEstate} from '../../actions/estate';
+
 import PropTypes from 'prop-types';
+import Prompt from '../Prompt/Prompt';
 
 const Estate = ({userId, role, estate, deleteEstate, location, history}) => {
+    const [prompt, setPrompt] = useState({
+        show: false,
+        confirm: false,
+        promptMsg: '',
+    });
+    const {show, confirm, promptMsg} = prompt;
+    useEffect(() => {
+        if (confirm) {
+            //location.pathname is a prop from withRouter which gives me current url eg./dashboard
+            deleteEstate(estate._id, location.pathname, history);
+        }
+    }, [confirm]);
     const handleDelete = () => {
-        //location.pathname is a prop from withRouter which gives me current url eg./dashboard
-        deleteEstate(estate._id, location.pathname, history);
+        setPrompt({...prompt, show: true, promptMsg: 'delete estate'});
+    };
+    const handleConfirm = () => {
+        setPrompt({...prompt, confirm: true});
+    };
+    const hidePrompt = () => {
+        setPrompt({...prompt, show: false});
     };
     let buttons;
     if (role === 'admin') {
@@ -54,8 +73,16 @@ const Estate = ({userId, role, estate, deleteEstate, location, history}) => {
             </div>
         );
     }
+
     return (
         <div className='estate-container'>
+            {show && (
+                <Prompt
+                    msg={promptMsg}
+                    toggleConfirm={handleConfirm}
+                    toggleShow={hidePrompt}
+                />
+            )}
             <h3 className='estate-name'>{estate.name}</h3>
             <div className='estate-description'>
                 <p className='estate-description_text'>{estate.description}</p>
